+++ packages/react-test-renderer/src/ReactShallowRendererEntry.js
