+++ packages/react-test-renderer/src/ReactTestRendererFiberEntry.js
